Refactor model definitions and streamline error handling
Extracted model structs into a new `models` module for improved modularity and maintenance.
